requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bcc==0.1.10
blinker==1.4
cachetools==6.2.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.3
chrome-gnome-shell==0.0.0
click==8.3.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.2.1
defer==1.0.6
distro==1.7.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
frozenlist==1.7.0
genai==2.1.0
google-auth==2.41.0
google-genai==1.39.1
gyp==0.1
h11==0.16.0
hidpidaemon==18.4.6
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
ipython==8.37.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.6.4
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
openai==0.27.10
parso==0.8.5
pexpect==4.9.0
pop-transition==1.1.2
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==3.12.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycups==2.0.1
pydantic==2.11.9
pydantic_core==2.33.2
pydbus==0.6.0
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.1
python-gnupg==0.4.8
python-multipart==0.0.20
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2025.9.18
repolib==2.2.1
repoman==1.4.0
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
SecretStorage==3.3.1
sentry-sdk==2.39.0
sessioninstaller==0.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.48.0
systemd-python==234
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
xdg==5
xkit==0.0.0
yarl==1.20.1
zipp==1.0.0
